requirements.txt
keras==2.0.5
tensorflow-gpu==1.1.0
cudnn==6.0
cudatoolkit==8.0